finalizing automatic bin shifting for correlation histograms

+++ src/corrFunctions.hh
@@ -221,8 +221,8 @@ namespace corrAnalysis {
     unsigned binsEta;             // Binning for histograms
     unsigned binsPhi;             // Binning for histograms
     
-    double phiBinShift;           // value for shifting histogram edges to force a bin centered at zero
-    double etaBinShift;           // value for shifting histogram edges to force a bin centered at zero
+    float phiBinShift;            // value for shifting histogram edges to force a bin centered at zero
+    float etaBinShift;            // value for shifting histogram edges to force a bin centered at zero
 		
 		// Event statistics
 		TH3D* hCentVz;
